move test files to src folder
