+++ packages/web3torrent/src/library/client-seeding.test.ts
@@ -3,13 +3,13 @@ import WebTorrentPaidStreamingClient, {
   ClientEvents,
   PaidStreamingExtensionNotices,
   PaidStreamingTorrent
-} from './../src/library/web3torrent-lib';
+} from './web3torrent-lib';
 import {
   defaultFile,
   defaultFileMagnetURI,
   defaultSeedingOptions,
   defaultTorrentHash
-} from './utils';
+} from './testing/test-utils';
 
 describe('Seeding and Leeching', () => {
   let seeder: WebTorrentPaidStreamingClient;
